Add files via upload
a work in progress. this version works and shows well in FF; chrome has issues displaying the font compression necessary for the cards to look like anything.

the idea is to create a deck of playable cards using only html, css and js. 
so far it works, but the project is in "deal-only mode", meaning no one can play with it yet. if you have suggestions or want to fork to add drag and drop functionality, multiplayer usability etc, by all means...